requirements.txt
pandas
matplotlib
IPython
numpy
scikit-learn
statsmodels
hydra-core